+++ contracts/utils/StringsUtil.sol
@@ -3,29 +3,28 @@
 pragma solidity ^0.8.0;
 
 library StringsUtil {
-    function length(string memory _base)
-    internal
-    pure
-    returns (uint) {
+    function length(string memory _base) internal pure returns (uint256) {
         bytes memory _baseBytes = bytes(_base);
         return _baseBytes.length;
     }
 
     function concat(string memory _base, string memory _value)
-    internal
-    pure
-    returns (string memory) {
+        internal
+        pure
+        returns (string memory)
+    {
         bytes memory _baseBytes = bytes(_base);
         bytes memory _valueBytes = bytes(_value);
 
         require(_valueBytes.length > 0, "Strings: empty value");
 
-        string memory _tmpValue = new string(_baseBytes.length +
-            _valueBytes.length);
+        string memory _tmpValue = new string(
+            _baseBytes.length + _valueBytes.length
+        );
         bytes memory _newValue = bytes(_tmpValue);
 
-        uint i;
-        uint j;
+        uint256 i;
+        uint256 j;
 
         for (i = 0; i < _baseBytes.length; i++) {
             _newValue[j++] = _baseBytes[i];
@@ -39,20 +38,20 @@ library StringsUtil {
     }
 
     function split(string memory _base, string memory _value)
-    internal
-    pure
-    returns (string[] memory splitArr) {
+        internal
+        pure
+        returns (string[] memory splitArr)
+    {
         bytes memory _baseBytes = bytes(_base);
 
-        uint _offset = 0;
-        uint _splitsCount = 1;
+        uint256 _offset = 0;
+        uint256 _splitsCount = 1;
         while (_offset < _baseBytes.length - 1) {
-            int _limit = _indexOf(_base, _value, _offset);
-            if (_limit == - 1)
-                break;
+            int256 _limit = _indexOf(_base, _value, _offset);
+            if (_limit == -1) break;
             else {
                 _splitsCount++;
-                _offset = uint(_limit) + 1;
+                _offset = uint256(_limit) + 1;
             }
         }
 
@@ -61,52 +60,52 @@ library StringsUtil {
         _offset = 0;
         _splitsCount = 0;
         while (_offset < _baseBytes.length - 1) {
-
-            int _limit = _indexOf(_base, _value, _offset);
-            if (_limit == - 1) {
-                _limit = int(_baseBytes.length);
+            int256 _limit = _indexOf(_base, _value, _offset);
+            if (_limit == -1) {
+                _limit = int256(_baseBytes.length);
             }
 
-            string memory _tmp = new string(uint(_limit) - _offset);
+            string memory _tmp = new string(uint256(_limit) - _offset);
             bytes memory _tmpBytes = bytes(_tmp);
 
-            uint j = 0;
-            for (uint i = _offset; i < uint(_limit); i++) {
+            uint256 j = 0;
+            for (uint256 i = _offset; i < uint256(_limit); i++) {
                 _tmpBytes[j++] = _baseBytes[i];
             }
-            _offset = uint(_limit) + 1;
+            _offset = uint256(_limit) + 1;
             splitArr[_splitsCount++] = string(_tmpBytes);
         }
         return splitArr;
     }
 
-    function _indexOf(string memory _base, string memory _value, uint _offset)
-    internal
-    pure
-    returns (int) {
+    function _indexOf(
+        string memory _base,
+        string memory _value,
+        uint256 _offset
+    ) internal pure returns (int256) {
         bytes memory _baseBytes = bytes(_base);
         bytes memory _valueBytes = bytes(_value);
 
         assert(_valueBytes.length == 1);
 
-        for (uint i = _offset; i < _baseBytes.length; i++) {
+        for (uint256 i = _offset; i < _baseBytes.length; i++) {
             if (_baseBytes[i] == _valueBytes[0]) {
-                return int(i);
+                return int256(i);
             }
         }
 
-        return - 1;
+        return -1;
     }
-//
-//    function replace(string memory _string, string memory _string, string memory _letter) internal pure returns (string memory) {
-//        bytes memory _stringBytes = bytes(_string);
-//        bytes memory result = new bytes(_stringBytes.length);
-//
-//        for (uint i = 0; i < _stringBytes.length; i++) {
-//            result[i] = _stringBytes[i];
-//            if (i == _pos)
-//                result[i] = bytes(_letter)[0];
-//        }
-//        return string(result);
-//    }
+    //
+    //    function replace(string memory _string, string memory _string, string memory _letter) internal pure returns (string memory) {
+    //        bytes memory _stringBytes = bytes(_string);
+    //        bytes memory result = new bytes(_stringBytes.length);
+    //
+    //        for (uint i = 0; i < _stringBytes.length; i++) {
+    //            result[i] = _stringBytes[i];
+    //            if (i == _pos)
+    //                result[i] = bytes(_letter)[0];
+    //        }
+    //        return string(result);
+    //    }
 }
